rnn: parse <copyright> tag
Information in the <copyright> tag is put in rnncopyright which directly
included in rnndb.

Authors from multiple tags are merged (but currently without trying to
avoid duplicate authors) and the earliest copyright year is taken.

However, if <license> is specified multiple times, the contents must
be identical, or a fatal error is issued.

This could be solved by either making licenses per-author or concatenating
the licenses; however, we shouldn't need to use this, it complicates things
and is of dubious usefulness and legal standing anyway.

It is however possible to specify an identical license blurb multiple times,
allowing the traditional license-blurb-in-every-file pattern.