demmt: remove write buffering
Write buffering was crucial in early demmt days, when my
understanding of how blob communicates with the GPU was much worse.
Over time write buffering turned out to annoy some people because
it hid the true order of events.

Now that demmt knows a lot more and this feature is not needed
anymore, we can remove it.

As a bonus it fixes some minor bugs.