envydis: gk110: Add support for dadd with an immediate src
This commit adds support for dadd with an immediate src in gk110 code.
The machine-code in question is generated by e.g. nouveau_compiler with
the new "Make use of double immediates" patch series when building the
piglit glsl-algebraic-double-add.shader_test.

This commit changes the output from:

00000010: 001c0001 c38001ff     $r0 $r0 $r0 $r0 0x3fe00 0x3fe00
  0x3fe0000000000000 0x3fe00000 0x0 0x3 ???
  [unknown: 00000000 c0800000] [unknown instruction]

Into:

00000010: 001c0001 c38001ff     add rn f64 $r0d $r0d 0x3fe0000000000000

The machine-code in question disassembles to the same using
nvdisasm and works properly on an actual gpu.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>