gm107: separate out sched decoding from regular ops
Unfortunately there's no sure-fire way to distinguish a sched "op" from
regular ones. They happen purely based on position. So for disassembly,
use the byte position in the file to determine whether the current op
should be decoded as sched or as regular.

For assembly, we go back to using a shared table. Ideally we would pick
out which one to use based on the position in the outputted bytestream,
but it was not apparent to me how to do that. Secondly, it would make it
impossible to assembly one plain instruction without first inserting a
sched, which would be annoying.

Note - this does mean that to decode a gm107 op, one has to supply sched
bytes first. However we're nowhere as picky as nvdisasm - zeroes work
just fine.