demmt: more gracefully handle GET_CHIPSET ioctl failure
This is a hack. Really it should be sufficient for things to call
nvrm_get_chipset. But in practice, they want to call other things like
nvrm_get_pb_pointer_found and others, which triggers null asserts. We
could go around trying to fix up all the places, but it seems pointless
to do so.

This makes traces with broken GET_CHIPSET ioctl's able to get decoded
with -m. It also points out the issue that if there is no GET_CHIPSET
*at all*, the -m won't save you. This can be fixed in a future commit,
and this hack of setting the chipset to itself can be removed.