rnndb: Add more missing registers for GK110+
These are all of the slcg registers I've found thus-far that weren't
already documented in rnndb that I've come across in the vbios repo. As
well, the sources I used to verify each as much as possible:

Confirmed in android tegra repos:
 - 0x17e050
 - 0x17ea98
 - 0x13cc04 (this also made it obvious that 0x13cc00 exists beyond the
   GK104)

Confirmed in nvidia nvgpu repos:
 - 0x13c824

Guesses based off nearby register writes:
 - 0x10f280

The value of all of these registers is consistently 0x0, however that
seems to match up perfectly with the production values specified in
android's tegra drivers and nvgpu, with the disable value being
non-zero (0x0 must be some sort of enable I suppose).