nvagetbios: Add autodetection
If we know enough to say that the user should try a different extraction
method, we should just save them the hassle and try a different
extraction method ourselves by trying PRAMIN, then PROM (and vice versa
for older cards). If we don't manage to get a valid vbios (whether it be
from signature verification failures or something else), abort so the
user can decide what to do.