nva: Fix pmc_id endianness check
So: this check is right for basically every card under the sun except
some rather weird ones. Turns out one of these weird ones is the
GK208B/NV108 (GT710) that I got in the mail, which appears to have a
PMC.ID value of 0xb060b0b1. This seems to be one of the rare cases
(which strangely is actually mentioned in the comments for nva already)
where both bits 31 and 7 are set to 1.

So: we know a couple of things now:
 - Pre-NV10: B31 and B7 are always 0
 - NV10+ LE: B31 is 0, B7 is 1 OR
             B31 is 1 and B7 is 1, and B8-11 are 0
 - NV10+ BE: B31 is 1, B7 is 0 OR
             B31 is 1 and B7 is 1, and B16-19 are 0

So, use this to come up with a better algorithm for detecting this
that's still valid across multiple generations. Now we can handle my
weird card :).