nvbios: Parse PCI Data Structure field 'Code Type'
PCI Data Structure offset 14h is `Code Type`, a one byte field that identifies
the type of code contained in this section of the ROM. The code may be executable
binary for a specific processor and system architecture or interpretive code.

The field has been present since at least BIOS Boot Specification, V1.01 (1996).

As of PCI Firmware Specification, Rev. 3.0 the following are valid values:

  00    Intel x86, PC-AT compatible
  01    Open Firmware standard for PCI (IEEE 1275-1994)
  02    Hewlett-Packard PA RISC
  03    Extensible Firmware Interface (EFI)
  04-FF Reserved