travis: Reflect Ubuntu Xenial packaging changes
libdrm_nouveau is no longer separately packaged in Ubuntu Xenial (16.04) (as it
was in Ubuntu Trusty (14.04)).

Reflect this in the Travis-CI configuration settings.

Use this opportunity to also check for the dependency on core libdrm itself, in
addition to libdrm built with nouveau support. This is a commonly used approach
in other CMake builds.