nvbios: Add INIT_I2C_WORD_CONDITION devinit opcode (0xB4)
Check the condition of a register with 16bit data over I2C port.

This opcode causes the indicated I2C condition to be tested.
If the condition is met the condition flag is unmodified.
If the condition is not met or an I2C read failure occurs the
condition flag is set to the skip operations state.

Seen on [core88,) e.g. Volta+.

Source:
- https://download.nvidia.com/open-gpu-doc/Devinit/1/devinit.xml#INIT_I2C_WORD_CONDITION
- https://nvidia.github.io/open-gpu-doc/Devinit/devinit.xml