nvbios: Document INIT_RESET_END devinit opcode (0x8D)
Signal that the reset sequence has completed.

This opcode signals that the software reset sequence has completed.
Ordinarily, no actual operations are performed by the opcode.
However it allows for possible software work arounds by devinit
engines in software agents other than the VBIOS, such as the resman,
FCODE, and EFI driver. This opcode is designed to be included in
the devinit script immediately after the NV_PMC_ENABLE register
has be written to take most of the engines out of the reset state.

Present on [core6,)

Source:
- https://download.nvidia.com/open-gpu-doc/Devinit/1/devinit.xml#INIT_RESET_END